Remove key unit flag from seek handler
The GStreamer docs recommended that playback applications use the
GST_SEEK_FLAG_KEY_UNIT mask in seeking operations, but I've noticed it
has the tendency to cause playback to resume a substantial amount before
or after the requested seek time in certain areas of certain files. A
quick smoke test seems to demonstrate that lacking this flag isn't
detrimental, so I've just removed it.