Improve user idle tracking
The old code to track whether the user was using the application or
simply watching was haphazard and ad-hoc. This commit adds extra junk
to Bus and moves the timeout code to PlaybackController, to centralise
the time constants and logic etc.

Also fixes a bug wherein a user seeking and playing after EOS returns
playback to the beginning of the file.